extract logging config into subcrate
